pyproject.toml
[project]
name = "api-bancaria-fastapi"
version = "0.1.0"
description = ""
authors = [{ name = "FilipeMadeira13", email = "cfilipemadeira@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlmodel",
    "alembic",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "bcrypt",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
